fix broken workflow for incorrect options address
Providing an incorrect paymentAddress in the workerOptions param would just break the hook and leave it in a broken state. There is a case where the param may not be defined on the first render/execution of the hook, but will be defined in the next render or a later time. The hook should be able to retry building the worker on subsequent rerenders.

Introduced the optionsError flag to capture when options validation and/or
construction fails before constructing the worker. Added handling of
optionsError to retry constructing the workerOptions obj and preventing worker intialization.

There were other React nature bugs that were resolved (useCallback
dependencies)

Also deprecated the options prop on workerStatistics since those values
can be found on the worker.