replace maxWorkingSandboxes with cores configuration
new supervisor in worker API deprecates the maxWorkingSandboxes concept
and replaces it with cores. maxWorkingSandboxes is still supported but
will eventually be completely deprecated. Similarily, this commit
supports the new cores property as maxWorkingSandboxes was supported,
while still keeping the latter

STILL WIP - as of testing, proxy logic seems broken (not saving cores to
local storage)