fix!: remove autosaving to payment address to local storage
It's now up to the application to handle persisting the payment address
of the worker. For some apps, like the dashboard app at
`https://dcp.cloud/`, it makes more sense to let the app persist the
config value as a user preference associated with a user's account,
rather than a "machine" preference.

BREAKING CHANGE: The `paymentAddress` is no longer saved in local storage.